New test was added- TokensTests.java

+++ src/main/java/biz/bsoft/users/dao/PasswordResetTokenRepository.java
@@ -11,7 +11,7 @@ import java.util.Date;
 /**
  * Created by vbabin on 16.08.2016.
  */
-public interface PasswordResetTokenRepository extends CrudRepository<PasswordResetToken, Long> {
+public interface PasswordResetTokenRepository extends CrudRepository<PasswordResetToken, Integer> {
     PasswordResetToken findByToken(String token);
 
     @Modifying
